board:2093 bugprone-narrowing-conversions reaches ZERO, and the atmosphere stops computing twice
  this check        59 -> 0
  findings        5419 -> 5360
  Venice      c713d304 -> 05b6863d
  OldTown     b0f9f7d7 -> a23e97f2

THE 38 IN MediumCore.h WERE A REAL DIVERGENCE, not a style complaint. That header is compiled TWICE
-- once as C++ and once as Metal -- and the two were not computing the same thing. MSL has no
double, so 1.0 there is a float; in C++ it is a double, and every intermediate widened. The C++
shims made it worse: max and clamp took double, so a float went up to double and came back. And
OUTSHINE_PI was std::numbers::pi on one side and a printed float on the other.

Both sides now use float throughout: the literals carry f, the shims take float, and the macro is
std::numbers::pi_v<float>. The same source finally means the same arithmetic in both compilations.
That moves the digests, on purpose, and Venice was looked at afterwards: lagoon, roofs and the
atmosphere gradient all stand.

THE CHECK ALSO FOUND A LYING DIAGNOSTIC. Document.cpp formats accessor bounds through
Number(size_t), so a message reporting that an element "runs -0.5 to 1.5 over a declared 0 to 1" was
printing the truncated integers instead. Nine call sites now use Decimal(double). A refusal that
misreports the numbers it refuses over is worse than no refusal.

And two of my own from earlier rounds: making kCoarseDrop unsigned turned Surface_.Z - kCoarseDrop
into an unsigned subtraction assigned to an int, and the uint64 shift results were assigned to long
without saying so. Both now say what they do.

+++ src/world/ground/ClassField.cpp
@@ -238,7 +238,7 @@ void ClassField::Update(TilePool &tiles, double camLat, double camLon, double bu
                {{"version", static_cast<double>(done->Structure->Version())},
                 {"buildMs", buildMs},
                 {"packMs", done->Structure->Measured().PackMs},
-                {"bufferKB", done->Structure->Bytes() / 1024.0},
+                {"bufferKB", static_cast<double>(done->Structure->Bytes()) / 1024.0},
                 {"lentKB",
                  static_cast<double>(CapacityBytes(t.Pts) + CapacityBytes(t.Rings) +
                                      CapacityBytes(t.Feats)) /
